@@ -15,16 +15,6 @@
 
 package com.amazonaws.geo.model;
 
-import com.amazonaws.services.dynamodbv2.model.GetItemResult;
+public abstract class GeoDataResponse {
 
-public class GetPointResult extends GeoDataResult {
-	private GetItemResult getItemResult;
-
-	public GetPointResult(GetItemResult getItemResult) {
-		this.getItemResult = getItemResult;
-	}
-
-	public GetItemResult getGetItemResult() {
-		return getItemResult;
-	}
 }
