@@ -15,10 +15,7 @@
 
 package com.amazonaws.geo.model;
 
-import java.util.HashMap;
-
-import com.amazonaws.services.dynamodbv2.model.AttributeValue;
-import com.amazonaws.services.dynamodbv2.model.GetItemRequest;
+import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
 
 /**
  * Get point request. The request must specify a geo point and a range key value. You can modify GetItemRequest to
@@ -26,20 +23,12 @@ import com.amazonaws.services.dynamodbv2.model.GetItemRequest;
  * attribute will be overwritten by GeoDataManagerConfiguration.
  * */
 public class GetPointRequest extends GeoDataRequest {
-	private GetItemRequest getItemRequest;
 	private GeoPoint geoPoint;
 	private AttributeValue rangeKeyValue;
 
 	public GetPointRequest(GeoPoint geoPoint, AttributeValue rangeKeyValue) {
-		getItemRequest = new GetItemRequest();
-		getItemRequest.setKey(new HashMap<String, AttributeValue>());
-
 		this.geoPoint = geoPoint;
 		this.rangeKeyValue = rangeKeyValue;
-	}
-
-	public GetItemRequest getGetItemRequest() {
-		return getItemRequest;
 	}
 
 	public GeoPoint getGeoPoint() {
