@@ -15,12 +15,8 @@
 
 package com.amazonaws.geo.model;
 
-import java.util.HashMap;
-
 import com.amazonaws.geo.GeoDataManagerConfiguration;
-import com.amazonaws.services.dynamodbv2.model.AttributeValue;
-import com.amazonaws.services.dynamodbv2.model.PutItemRequest;
-import com.amazonaws.services.dynamodbv2.model.PutRequest;
+import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
 
 /**
  * Put point request. The request must specify a geo point and a range key value. You can modify PutItemRequest to
@@ -31,23 +27,12 @@ import com.amazonaws.services.dynamodbv2.model.PutRequest;
  * 
  * */
 public class PutPointRequest extends GeoDataRequest {
-	private PutItemRequest putItemRequest;
-	private PutRequest putRequest;
 	private GeoPoint geoPoint;
 	private AttributeValue rangeKeyValue;
 
 	public PutPointRequest(GeoPoint geoPoint, AttributeValue rangeKeyValue) {
-		putItemRequest = new PutItemRequest();
-		putItemRequest.setItem(new HashMap<String, AttributeValue>());
-		putRequest = new PutRequest();
-		putRequest.setItem(new HashMap<String, AttributeValue>());
-		
 		this.geoPoint = geoPoint;
 		this.rangeKeyValue = rangeKeyValue;
-	}
-
-	public PutItemRequest getPutItemRequest() {
-		return putItemRequest;
 	}
 
 	public GeoPoint getGeoPoint() {
@@ -56,9 +41,5 @@ public class PutPointRequest extends GeoDataRequest {
 
 	public AttributeValue getRangeKeyValue() {
 		return rangeKeyValue;
-	}
-	
-	public PutRequest getPutRequest() {
-		return putRequest;
 	}
 }
