@@ -15,16 +15,9 @@
 
 package com.amazonaws.geo.model;
 
-import com.amazonaws.services.dynamodbv2.model.PutItemResult;
+public class QueryRadiusResponse extends GeoQueryResponse {
 
-public class PutPointResult extends GeoDataResult {
-	private PutItemResult putItemResult;
-
-	public PutPointResult(PutItemResult putItemResult) {
-		this.putItemResult = putItemResult;
-	}
-
-	public PutItemResult getPutItemResult() {
-		return putItemResult;
+	public QueryRadiusResponse(GeoQueryResponse geoQueryResponse) {
+		super(geoQueryResponse);
 	}
 }
