@@ -17,10 +17,9 @@ package com.amazonaws.geo.util;
 
 import java.io.IOException;
 
-import org.codehaus.jackson.map.ObjectMapper;
-
 import com.amazonaws.geo.model.GeoObject;
 import com.amazonaws.geo.model.GeoPoint;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 public class GeoJsonMapper {
 	private static ObjectMapper mapper = new ObjectMapper();
