@@ -16,15 +16,15 @@
 package com.amazonaws.geo.util;
 
 import com.amazonaws.geo.GeoDataManagerConfiguration;
-import com.amazonaws.services.dynamodbv2.model.AttributeDefinition;
-import com.amazonaws.services.dynamodbv2.model.CreateTableRequest;
-import com.amazonaws.services.dynamodbv2.model.KeySchemaElement;
-import com.amazonaws.services.dynamodbv2.model.KeyType;
-import com.amazonaws.services.dynamodbv2.model.LocalSecondaryIndex;
-import com.amazonaws.services.dynamodbv2.model.Projection;
-import com.amazonaws.services.dynamodbv2.model.ProjectionType;
-import com.amazonaws.services.dynamodbv2.model.ProvisionedThroughput;
-import com.amazonaws.services.dynamodbv2.model.ScalarAttributeType;
+import software.amazon.awssdk.services.dynamodb.model.AttributeDefinition;
+import software.amazon.awssdk.services.dynamodb.model.CreateTableRequest;
+import software.amazon.awssdk.services.dynamodb.model.KeySchemaElement;
+import software.amazon.awssdk.services.dynamodb.model.KeyType;
+import software.amazon.awssdk.services.dynamodb.model.LocalSecondaryIndex;
+import software.amazon.awssdk.services.dynamodb.model.Projection;
+import software.amazon.awssdk.services.dynamodb.model.ProjectionType;
+import software.amazon.awssdk.services.dynamodb.model.ProvisionedThroughput;
+import software.amazon.awssdk.services.dynamodb.model.ScalarAttributeType;
 
 /**
  * Utility class.
@@ -50,31 +50,31 @@ public class GeoTableUtil {
 	 * @return Generated create table request.
 	 */
 	public static CreateTableRequest getCreateTableRequest(GeoDataManagerConfiguration config) {
-		CreateTableRequest createTableRequest = new CreateTableRequest()
-				.withTableName(config.getTableName())
-				.withProvisionedThroughput(
-						new ProvisionedThroughput().withReadCapacityUnits(10L).withWriteCapacityUnits(5L))
-				.withKeySchema(
-						new KeySchemaElement().withKeyType(KeyType.HASH).withAttributeName(
-								config.getHashKeyAttributeName()),
-						new KeySchemaElement().withKeyType(KeyType.RANGE).withAttributeName(
-								config.getRangeKeyAttributeName()))
-				.withAttributeDefinitions(
-						new AttributeDefinition().withAttributeType(ScalarAttributeType.N).withAttributeName(
-								config.getHashKeyAttributeName()),
-						new AttributeDefinition().withAttributeType(ScalarAttributeType.S).withAttributeName(
-								config.getRangeKeyAttributeName()),
-						new AttributeDefinition().withAttributeType(ScalarAttributeType.N).withAttributeName(
-								config.getGeohashAttributeName()))
-				.withLocalSecondaryIndexes(
-						new LocalSecondaryIndex()
-								.withIndexName(config.getGeohashIndexName())
-								.withKeySchema(
-										new KeySchemaElement().withKeyType(KeyType.HASH).withAttributeName(
-												config.getHashKeyAttributeName()),
-										new KeySchemaElement().withKeyType(KeyType.RANGE).withAttributeName(
-												config.getGeohashAttributeName()))
-								.withProjection(new Projection().withProjectionType(ProjectionType.ALL)));
+		CreateTableRequest createTableRequest = CreateTableRequest.builder()
+				.tableName(config.getTableName())
+				.provisionedThroughput(
+						ProvisionedThroughput.builder().readCapacityUnits(10L).writeCapacityUnits(5L).build())
+				.keySchema(
+						KeySchemaElement.builder().keyType(KeyType.HASH).attributeName(
+								config.getHashKeyAttributeName()).build(),
+						KeySchemaElement.builder().keyType(KeyType.RANGE).attributeName(
+								config.getRangeKeyAttributeName()).build())
+				.attributeDefinitions(
+						AttributeDefinition.builder().attributeType(ScalarAttributeType.N).attributeName(
+								config.getHashKeyAttributeName()).build(),
+						AttributeDefinition.builder().attributeType(ScalarAttributeType.S).attributeName(
+								config.getRangeKeyAttributeName()).build(),
+						AttributeDefinition.builder().attributeType(ScalarAttributeType.N).attributeName(
+								config.getGeohashAttributeName()).build())
+				.localSecondaryIndexes(
+						LocalSecondaryIndex.builder()
+								.indexName(config.getGeohashIndexName())
+								.keySchema(
+										KeySchemaElement.builder().keyType(KeyType.HASH).attributeName(
+												config.getHashKeyAttributeName()).build(),
+										KeySchemaElement.builder().keyType(KeyType.RANGE).attributeName(
+												config.getGeohashAttributeName()).build())
+								.projection(Projection.builder().projectionType(ProjectionType.ALL).build()).build()).build();
 
 		return createTableRequest;
 	}
